Add support for new webpack asset manifest format
webpack-asset-manifest introduced a new format in v4. In order to
support prefetch and preload, assets are not nested under an assets key
in the entrypoints object. This commit updates the manifest parsing
logic to support this new version in addition to the previous.

More info on the manifest changes can be found at
https://www.npmjs.com/package/webpack-assets-manifest#new-in-version-4